+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## XX.XX.XX
-* Added a config method to disable server config in the initialization "disableSDKBehaviorSettings()".
+* Added a config method to disable server config in the initialization "disableSDKBehaviorSettingsUpdates()".
 
 ## 25.4.1
 * Mitigated an issue that could occur while serializing events to improve stability, performance and memory usage.